% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_end.R
\name{Intrinsic_MRF_Prec}
\alias{Intrinsic_MRF_Prec}
\title{Create Markov Random Field Precision matrix}
\usage{
Intrinsic_MRF_Prec(knots_r1, tau)
}
\arguments{
\item{knots_r1}{1d vector or 2d matrix of knot locations.}

\item{precision}{parameter, positive constant.}
}
\value{
nrow(knots_r1) square matrix with intrinsic MRF matrix.
}
\description{
Create Markov Random Field Precision matrix
}
\examples{
Intrinsic_MRF_Prec(c(1,2,3),2)
}
