% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fn_optim.R
\name{mkl_fn_seq_em_tau_pred}
\alias{mkl_fn_seq_em_tau_pred}
\title{Predictions and summaries from msss_fit}
\usage{
mkl_fn_seq_em_tau_pred(locations, results, design_mat = NULL,
  model_used = 1, type = "pred")
}
\arguments{
\item{locations}{1d vector or 2d matrix of locations of the locations you'd like to predict at}

\item{results}{output from mr_optim_fit}

\item{design_mat}{if fixed effects are part of the model, design matrix for the locations}

\item{model_used}{which of the models to use, less than or equal to results$total_iters}

\item{type}{'pred' is default for prediction interval, 'mean' is interval for the mean, 'noint' is just a prediction without lower and upper bounds and is much faster, 'rescount' counts how many resoltuions are active at each point}
}
\value{
list with either a row matrix of predictions or a row matrix of resolutions active
FILLER
}
\description{
Predictions and summaries from msss_fit
}
