% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_end_optim.R
\name{mr_optim_fit}
\alias{mr_optim_fit}
\title{Runs multiresolution optimization routine}
\usage{
mr_optim_fit(yy, locationss, knots_r1, spatial_dimension, kernel_width,
  smoothness, shrinkage, maxiter, lambdatree_seq, tau_init = tau_init,
  tau_a = tau_a, tau_b = tau_b, em_tol = 0.001, m_tol = 1e-05,
  design_mat = NULL)
}
\arguments{
\item{yy}{response values}

\item{locationss}{locations of observations, either 1d vector or matrix with 2 columns}

\item{spatial_dimension}{spatial dimension of observations, either 1 or 2}

\item{kernel_width}{width of the kernel}

\item{smoothness}{smoothness parameter for kernel}

\item{shrinkage}{vector of parameter d, the shrinkage that increases in resolution, long vector}

\item{maxiter}{maximum number of iterations to run}

\item{lambdatree_seq}{decreasing sequence of lambdas, the parameter that controls sparsity}

\item{tau_init}{initial tau value}

\item{tau_a}{a parameter for gamma prior on tau}

\item{tau_b}{b parameter for gamma prior on tau}

\item{em_tol}{tolerance for when tau has converged}

\item{m_tol}{default is .00001, tolerance for when the coefficients have converged}

\item{design_mat}{optional design matrix for fixed effects}
}
\value{
list to be processed by prediction_function
}
\description{
Runs multiresolution optimization routine
}
