% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fn_optim.R
\name{mr_optim_pred}
\alias{mr_optim_pred}
\title{Predictions and summaries from mr_optim_fit}
\usage{
mr_optim_pred(locations, results, design_mat = NULL,
  marg_type = "bic_CAP", int_type = "pred", level = 0.95,
  a_beta = 1, b_beta = 1, a_g = 3)
}
\arguments{
\item{locations}{1d vector or 2d matrix of locations of the locations you'd like to predict at}

\item{results}{output from mr_optim_fit}

\item{design_mat}{if fixed effects are part of the model, design matrix for the locations}

\item{marg_type}{howt to calculate the marginal, "bic_CAP" is default and strongly recommended but "hyperg_refit" is an option}

\item{int_type}{"pred" for prediction, "recount" to count resolutions, "noint" for a prediction only}

\item{level}{confidence level for the interval}

\item{a_beta}{a parameter for beta prior on model space, 1 is default}

\item{b_beta}{b parameter for beta prior on model space, 1 is default, but if data is large, use 10^(3*n/10^4)}

\item{a_g}{if "hyperg_refit" is used}
}
\value{
list with either a row matrix of predictions or a row matrix of resolutions active
FILLER
}
\description{
Predictions and summaries from mr_optim_fit
}
