\name{MSSS-package}
\alias{MSSS-package}
\alias{MSSS}
\docType{package}
\title{
\packageTitle{MSSS}
}
\description{
\packageDescription{MSSS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MSSS}
\packageIndices{MSSS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MSSS}

Maintainer: \packageMaintainer{MSSS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
