\docType{package}
\name{MSSS}
\alias{MSSS}
\alias{MSSS-package}
\title{MSSS: Multiresolution Shotgun Stochastic Search}
\description{
Package for fitting spatial models with multiresolution Kernel convolutions and spatially varying resolution.
}

\references{
Kirsner, Daniel, and Bruno Sansó. "Multi-scale shotgun stochastic search for large spatial datasets." Computational Statistics & Data Analysis 146 (2020): 106931.  \cr
\url{https://www.sciencedirect.com/science/article/abs/pii/S0167947320300220}
}
\seealso{

\code{\link{msss_fit}()},
\code{\link{msss_pred}()},
\code{\link{Intrinsic_MRF_Prec}()},
\code{\link{r1_create}()}
}
\author{
Daniel Kirsner, \cr Maintainer: Daniel Kirsner <dkirsner@ucsc.edu>
}
\keyword{package}
\keyword{spatial}