% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/front_end.R
\name{msss_fit}
\alias{msss_fit}
\title{Create R1 knots on a grid a specificed number of knots wide}
\usage{
msss_fit(locations, yy, knots_r1, spatial_dimension, maxiters = 100,
  cores = 1, design_mat = NULL, stopping_rule = "maxiter",
  g_method = 1, a_pi = 1, b_pi = 5, a_g = 3, kernel_width = 1.5,
  nu = 1, pi_method = 1, R1_prior = NULL)
}
\arguments{
\item{locations}{locations of spatial observations.}

\item{yy}{observation at each location}

\item{knots_r1}{locations of R1 knots, equally spaced grid, output of r1_create}

\item{spatial_dimension}{dimension of the locations, should be 1 or 2}

\item{maxiters}{number of iterations, default is 100}

\item{cores}{number of cores for parallel processing, ~20 is optimal}

\item{design_mat}{fixed effects, at least an intercept is recommended}

\item{stopping_rule}{'maxiter' is default, could be 'log likelihood' if you want it to exit early when finding no new models}

\item{g_method}{1 for hyper g, 1 i sdefault}

\item{a_pi}{parameter 1 for beta binomial sparsity if pi_method=1, or numerator if pi_method=2, 1 is default}

\item{b_pi}{parameter 2 for beta binomial sparsity if pi_method=1, or denominator if pi_method=2, 5 is default}

\item{a_g}{hyper g prior parameter, 3 is default, 2-4 recommended}

\item{kernel_width}{bezier kernel width parameter, should be 1.5 or greater to be sensible, 1.5 is default}

\item{nu}{bezier smoothness parameter, default is 1}

\item{pi_method}{beta binomial is 1 (default), binomial (fixed pi) is 2}

\item{R1_prior}{covariance matrix for R1 knots, default is NULL}
}
\value{
large list, exists to be processed by the function msss_predict forsummaries and predictions
}
\description{
Create R1 knots on a grid a specificed number of knots wide
}
\examples{
blank
}
