% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_fn.R
\name{msss_pred}
\alias{msss_pred}
\title{Predictions and summaries from msss_fit}
\usage{
msss_pred(locations, results, design_mat = NULL, level = 0.95,
  model_used = 1:100, type = "pred")
}
\arguments{
\item{locations}{1d vector or 2d matrix of locations of the locations you'd like to predict at}

\item{results}{output from msss_fit}

\item{design_mat}{if fixed effects are part of the model, design matrix for the locations}

\item{level}{confidence level, not always needed, .95 is defaykt}

\item{model_used}{vector of how many top models to use for Bayesian model averaging, 1:100 is default, if slow try 1:10}

\item{type}{'pred' is default for prediction interval, 'mean' is interval for the mean, 'noint' is just a prediction without lower and upper bounds and is much faster, 'rescount' counts how many resoltuions are active at each point}
}
\value{
list with either a row matrix of predictions, a row matrix of resolutions active, or 3 row matricies with upper bounds, predicted values, and lower bound
}
\description{
Predictions and summaries from msss_fit
}
\examples{
FILLER
}
