\name{MSSS-package}
\alias{MSSS-package}
\alias{MSSS}
\docType{package}
\title{
\packageTitle{MSSS}
}
\description{
\packageDescription{MSSS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MSSS}
\packageIndices{MSSS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MSSS}

Maintainer: \packageMaintainer{MSSS}
}
\references{
Kirsner, Daniel, and Bruno Sansó. "Multi-scale shotgun stochastic search for large spatial datasets." Computational Statistics & Data Analysis 146 (2020): 106931
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\examples{
#filler
}
